spirit-tokio: Either
Similar to the one from the either crate or futures::future::Either. But
this one implements traits relevant to us ‒ namely, untagged Deserialize
for parsing alternative configurations, our own traits and the future's
ones (not only Future, but Sink and Stream too).

This should make it possible to accept configuration that is alternative
between eg. TcpListener and UnixListener and plug the same function into
`per_connection` into it.