on_terminate takes FnOnce instead of FnMut
It is called just once anyway, so it is more correct and less limiting.

This is *technically* an API breaking change, since someone could have
implemented the `Extensible` trait for their own type. But it is
unlikely to have happened, the fix is small enough and furthermore, we
are pre-1 currently, so we probably can afford to bump just patch
version number.