Provide Extensible::around_hooks
Similar to how Extensible::run_around can provide context for the actual
application started within Builder::run, this one provides a similar
service for the configuration and other callbacks/hooks.

The motivation is actually similar ‒ being able to set up access to the
Tokio runtime (or similar thing) even for the pipelines that run in the
background thread.